+++ project-base/src/Shopsys/ShopBundle/Model/Transport/Transport.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Transport;
+
+use Doctrine\ORM\Mapping as ORM;
+use Shopsys\FrameworkBundle\Model\Transport\Transport as BaseTransport;
+use Shopsys\FrameworkBundle\Model\Transport\TransportData as BaseTransportData;
+
+/**
+ * @ORM\Table(name="transports")
+ * @ORM\Entity
+ */
+class Transport extends BaseTransport
+{
+    /**
+     * @param \Shopsys\ShopBundle\Model\Transport\TransportData $transportData
+     */
+    public function __construct(BaseTransportData $transportData)
+    {
+        parent::__construct($transportData);
+    }
+
+    /**
+     * @param \Shopsys\ShopBundle\Model\Transport\TransportData $transportData
+     */
+    public function edit(BaseTransportData $transportData)
+    {
+        parent::edit($transportData);
+    }
+}