+++ project-base/src/Shopsys/ShopBundle/Model/Transport/TransportData.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Transport;
+
+use Shopsys\FrameworkBundle\Model\Transport\TransportData as BaseTransportData;
+
+class TransportData extends BaseTransportData
+{
+    public function __construct()
+    {
+        parent::__construct();
+    }
+}