+++ project-base/tests/ShopBundle/Database/Model/Order/OrderTransportAndPaymentTest.php
@@ -7,9 +7,9 @@ use Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactory;
 use Shopsys\FrameworkBundle\Model\Payment\PaymentFacade;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\Vat;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\VatData;
-use Shopsys\FrameworkBundle\Model\Transport\Transport;
-use Shopsys\FrameworkBundle\Model\Transport\TransportDataFactory;
+use Shopsys\FrameworkBundle\Model\Transport\TransportDataFactoryInterface;
 use Shopsys\FrameworkBundle\Model\Transport\TransportFacade;
+use Shopsys\ShopBundle\Model\Transport\Transport;
 use Tests\ShopBundle\Test\DatabaseTestCase;
 
 class OrderTransportAndPaymentTest extends DatabaseTestCase
@@ -395,7 +395,7 @@ class OrderTransportAndPaymentTest extends DatabaseTestCase
      * @param \Shopsys\FrameworkBundle\Model\Pricing\Vat\Vat $vat
      * @param bool[] $enabledForDomains
      * @param bool $hidden
-     * @return \Shopsys\FrameworkBundle\Model\Transport\Transport
+     * @return \Shopsys\ShopBundle\Model\Transport\Transport
      */
     public function getDefaultTransport(Vat $vat, $enabledForDomains, $hidden)
     {
@@ -434,10 +434,10 @@ class OrderTransportAndPaymentTest extends DatabaseTestCase
     }
 
     /**
-     * @return \Shopsys\FrameworkBundle\Model\Transport\TransportDataFactory
+     * @return \Shopsys\ShopBundle\Model\Transport\TransportDataFactory
      */
     public function getTransportDataFactory()
     {
-        return $this->getContainer()->get(TransportDataFactory::class);
+        return $this->getContainer()->get(TransportDataFactoryInterface::class);
     }
 }