+++ project-base/tests/ShopBundle/Database/Model/Vat/VatFacadeTest.php
@@ -9,7 +9,7 @@ use Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactory;
 use Shopsys\FrameworkBundle\Model\Payment\PaymentFacade;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\VatData;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\VatFacade;
-use Shopsys\FrameworkBundle\Model\Transport\TransportDataFactory;
+use Shopsys\FrameworkBundle\Model\Transport\TransportDataFactoryInterface;
 use Shopsys\FrameworkBundle\Model\Transport\TransportFacade;
 use Tests\ShopBundle\Test\DatabaseTestCase;
 
@@ -22,8 +22,8 @@ class VatFacadeTest extends DatabaseTestCase
         /* @var $vatFacade \Shopsys\FrameworkBundle\Model\Pricing\Vat\VatFacade */
         $transportFacade = $this->getContainer()->get(TransportFacade::class);
         /* @var $transportFacade \Shopsys\FrameworkBundle\Model\Transport\TransportFacade */
-        $transportDataFactory = $this->getContainer()->get(TransportDataFactory::class);
-        /* @var $transportDataFactory \Shopsys\FrameworkBundle\Model\Transport\TransportDataFactory */
+        $transportDataFactory = $this->getContainer()->get(TransportDataFactoryInterface::class);
+        /* @var $transportDataFactory \Shopsys\ShopBundle\Model\Transport\TransportDataFactory */
         $paymentDataFactory = $this->getContainer()->get(PaymentDataFactory::class);
         /* @var $paymentDataFactory \Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactory */
         $paymentFacade = $this->getContainer()->get(PaymentFacade::class);
@@ -36,7 +36,7 @@ class VatFacadeTest extends DatabaseTestCase
         $vatToReplaceWith = $this->getReference(VatDataFixture::VAT_HIGH);
         /* @var $vatToReplaceWith \Shopsys\FrameworkBundle\Model\Pricing\Vat\Vat */
         $transport = $this->getReference(TransportDataFixture::TRANSPORT_PERSONAL);
-        /* @var $transport \Shopsys\FrameworkBundle\Model\Transport\Transport */
+        /* @var $transport \Shopsys\ShopBundle\Model\Transport\Transport */
         $transportData = $transportDataFactory->createFromTransport($transport);
         $payment = $this->getReference(PaymentDataFixture::PAYMENT_CASH);
         /* @var $payment \Shopsys\FrameworkBundle\Model\Payment\Payment */